requirements.txt
numpy>=1.19.4
opencv-python>=4.9.0
tensorflow==2.9.1
keras==2.9.0
optuna>=2.3.0
aio_pika>=6.7.1
aiohttp>=3.7.3
dataclasses-json>=0.5.2
pandas>=1.2.0
pillow>=8.1.0
matplotlib==3.5.2
scikit-learn